requirements.txt
fastapi==0.96.0
motor==3.5.1
pymongo==4.8.0
python-dotenv==1.0.0
uvicorn==0.22.0